Allow specifying the name of the production env
Change bootstrap.sh to filter to the value of KZBEnv rather than the
hardcoded PRODUCTION. This allows the user to set the name of the
environment.

I needed this to be able to build a release build that was pointed at
our UAT environment for pen testing. I wanted to have the environments
filtered to UAT just as they would with a production build.